requirement.txt
autopep8
pylint
black
crypto
requests
mysqlclient
django-bootstrap-v5
django-bootstrap5
django-bootstrap
django-widget-tweaks
dotenv
Pillow